requirements.txt
PyYAML~=5.3.1
setuptools~=46.1.3